Using the new approach in classical
