+++ openquake/qa_tests_data/classical/case_19/job.ini
@@ -1,7 +1,7 @@
 [general]
 
 description = SHARE OpenQuake Computational Settings
-calculation_mode = classical_simple
+calculation_mode = classical
 random_seed = 23
 
 [geometry]
